feat: refactor classes into their own file
