+++ Functions/RIPA.Functions.Submission/Models/v1/Acttak.cs
@@ -1,0 +1,7 @@
+﻿namespace RIPA.Functions.Submission.Models.v1;
+
+public class Acttak
+{
+    public string Act_CD { get; set; }
+    public string Is_Con { get; set; }
+}
