+++ Functions/RIPA.Functions.Submission/Models/v1/Listacttak.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listacttak
+{
+    public List<Acttak> ActTak { get; set; }
+}
