+++ Functions/RIPA.Functions.Submission/Models/v1/Listbassearch.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listbassearch
+{
+    public List<string> BasSearch { get; set; }
+}
