+++ Functions/RIPA.Functions.Submission/Models/v1/Listbasseiz.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listbasseiz
+{
+    public List<string> BasSeiz { get; set; }
+}
