+++ Functions/RIPA.Functions.Submission/Models/v1/Listcb.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listcb
+{
+    public List<string> Cb { get; set; }
+}
