+++ Functions/RIPA.Functions.Submission/Models/v1/Listdisb.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listdisb
+{
+    public List<string> Disb { get; set; }
+}
