+++ Functions/RIPA.Functions.Submission/Models/v1/Listethn.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listethn
+{
+    public List<string> Ethn { get; set; }
+}
