+++ Functions/RIPA.Functions.Submission/Models/v1/Listperson_Stopped.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listperson_Stopped
+{
+    public List<Person_Stopped> Person_Stopped { get; set; }
+}
