+++ Functions/RIPA.Functions.Submission/Models/v1/Listproptype.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listproptype
+{
+    public List<string> PropType { get; set; }
+}
