+++ Functions/RIPA.Functions.Submission/Models/v1/Listresult.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listresult
+{
+    public List<Result> Result { get; set; }
+}
