+++ Functions/RIPA.Functions.Submission/Models/v1/Listsusp_T.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Listsusp_T
+{
+    public List<string> Susp_T { get; set; }
+}
