+++ Functions/RIPA.Functions.Submission/Models/v1/Location.cs
@@ -1,0 +1,9 @@
+﻿namespace RIPA.Functions.Submission.Models.v1;
+
+public class Location
+{
+    public string Loc { get; set; }
+    public string City { get; set; }
+    public string K12_Flag { get; set; }
+    public string K12Code { get; set; }
+}
