+++ Functions/RIPA.Functions.Submission/Models/v1/Officer.cs
@@ -1,0 +1,10 @@
+﻿namespace RIPA.Functions.Submission.Models.v1;
+
+public class Officer
+{
+    public string UID { get; set; }
+    public string Proxy { get; set; }
+    public string ExpYears { get; set; }
+    public string AT { get; set; }
+    public string ATOth { get; set; }
+}
