+++ Functions/RIPA.Functions.Submission/Models/v1/Perc.cs
@@ -1,0 +1,12 @@
+﻿namespace RIPA.Functions.Submission.Models.v1;
+
+public class Perc
+{
+    public Listethn ListEthn { get; set; }
+    public string Age { get; set; }
+    public string Is_LimEng { get; set; }
+    public Listdisb ListDisb { get; set; }
+    public string Gend { get; set; }
+    public string GendNC { get; set; }
+    public string LGBT { get; set; }
+}
