+++ Functions/RIPA.Functions.Submission/Models/v1/Primaryreason.cs
@@ -1,0 +1,13 @@
+﻿namespace RIPA.Functions.Submission.Models.v1;
+
+public class Primaryreason
+{
+    public string StReas { get; set; }
+    public string StReas_N { get; set; }
+    public string Tr_ID { get; set; }
+    public string Tr_O_CD { get; set; }
+    public string EDU_sec_CD { get; set; }
+    public string EDU_subDiv_CD { get; set; }
+    public Listsusp_T ListSusp_T { get; set; }
+    public string Susp_O_CD { get; set; }
+}
