+++ Functions/RIPA.Functions.Submission/Models/v1/Result.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace RIPA.Functions.Submission.Models.v1;
+
+public class Result
+{
+    public string ResCD { get; set; }
+    public List<string> Res_O_CD { get; set; }
+}
