+++ Functions/RIPA.Functions.Submission/Models/v2/DojStop.cs
@@ -1,0 +1,5 @@
+﻿namespace RIPA.Functions.Submission.Models.v2;
+
+public class DojStop
+{
+}
